Autorefactor out CursorResourceManager
